+++ packages/fighting-design/main/src/main.vue
@@ -1,17 +1,30 @@
-<script lang="ts" setup name="FMain">
+<script lang="ts" setup>
   import { Props } from './main'
+  import { computed } from 'vue'
+  import { isString } from '../../_utils'
+  import type { CSSProperties, ComputedRef } from 'vue'
 
-  defineProps(Props)
+  const prop = defineProps(Props)
+
+  const styleList: ComputedRef<CSSProperties> = computed((): CSSProperties => {
+    const { padding } = prop
+
+    return {
+      padding: isString(padding) ? padding : padding + 'px'
+    } as const
+  })
 </script>
 
 <script lang="ts">
-  export default {
+  import { defineComponent } from 'vue'
+
+  export default defineComponent({
     name: 'FMain'
-  }
+  })
 </script>
 
 <template>
-  <main class="f-main" :style="{ padding }">
+  <main class="f-main" :style="styleList">
     <slot />
   </main>
 </template>