+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - `f-header` 组件的 `height` 属性支持 `number` 类型
 - `f-footer` 组件新增 `padding` 类型
 - `f-footer` 组件的 `height` 属性支持 `number` 类型
+- `f-main` 组件的 `padding` 属性支持 `number` 类型
 
 ## 0.6.5-alpha.2 (2022-09-21)